collapse highlightStencil classes into one
